requirements.txt
Pillow==2.8.1